nat: add XDP NAT Gateway
XDP Gateway parses incoming packets, and modifies the IPv4 header
and UDP port numbers to the desired NAT.

The egress NAT is controlled using rules specified within a eBPF map.
The ingress packets daddr and dport is used to identify the rule.
The rule specifies the egress saddr, daddr and dport, whilst the
sport is chosen using a random number generator.
The generated egress NAT is checked within the existing flow map
to ensure it is a unique tuple. If it is unique, the new flows
are stored within the map for use by subsequent packets.

Aggressive Reap is also included to allow for the eBPF program to auto
delete the flow from the map, once the return ping response is received.
Aggressive reap can be enabled for certain ports that only accept ping
type flows.

Signed-off-by: David Beckett <david.beckett@netronome.com>
Reviewed-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Reviewed-by: Quentin Monnet <quentin.monnet@netronome.com>